# AP_Param: Revert "check for duplicate names when adding scripting params"