This reverts commit a562b113ee20e1a79b5f1bbbccd9d06e0a75b672 from PR #30523 .

This has severe performance impacts on both SITL/autotest and real hardware. In SITL it increases the (virtual?) time required for the `plane_aerobatics.lua` script to load from <1ms to ~800ms, causing the AerobaticsScripting test to fail as the script is not always loaded by the time its parameters are referenced. On CubeOrange, it increases the time required from ~850ms to ~2000ms; the original test might have been flawed.

The culprit appears to be the `find_by_name` function call added for each parameter. Its operation is inscrutable to Peter Barker and I, and removing it fixes the time issue.

Experimentation reveals that the `find` function call used by the existing `Parameter` system is much faster. A custom version patched to return the required `i` and `group_element` values for the test increases script load time to ~900ms.

Considering the potential issues posed by `find_by_name` and our inability to immediately devote time to fixes, it seems prudent to revert to get CI back on its feet and try to get the feature in later after fixing the parameter system or re-designing the test to use the faster function.

Removing the parameter table setup infrastructure from the aerobatics script results in a load time of ~65ms on hardware, illustrating that there are still severe inefficiencies in the parameter setup, but this is a pre-existing issue.