@IamPete1 2025-07-14
I have done a PR to switch to use `find`, it just does a pointer comparison on the parameter itself, its much nicer than messing with the token stuff. I should have thought of doing it this way the first time. https://github.com/ArduPilot/ardupilot/pull/30645